Publish stub
I'm stuck at the problem stated in the README: the output PDF
is too small and blurry. More controls are needed for the page
size but I can't seem to be able to do that without blurring
even more or using absurd amounts of memory in convert.

I've heard that there's a bag of gummi bears in it for whoever
solves this.